refactor: 검증 이벤트 제거 & OrderTable -> Order 패키지 참조 인터페이스로 의존성 역전 리팩토링
