+++ src/test/java/kitchenpos/order/domain/OrderValidatorTest.java
@@ -1,10 +1,8 @@
-package kitchenpos.ordertable.application;
+package kitchenpos.order.domain;
 
 import static kitchenpos.common.fixtures.OrderTableFixtures.ORDER_TABLE1_NUMBER_OF_GUESTS;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import kitchenpos.common.ServiceTest;
-import kitchenpos.order.application.event.OrderCreateValidationEvent;
 import kitchenpos.order.exception.OrderException;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.domain.OrderTableRepository;
@@ -13,28 +11,29 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
 
-class OrderTableEventListenerTest extends ServiceTest {
+@SpringBootTest
+class OrderValidatorTest {
 
     @Autowired
-    private OrderTableEventListener orderTableEventListener;
+    private OrderValidator orderValidator;
 
     @Autowired
     private OrderTableRepository orderTableRepository;
 
     @Nested
-    @DisplayName("OrderCreateEvent가 발행되어 validateOrderTable인 이벤트 리슨 로직이 실행될 때")
-    class ExecuteEventListener {
+    @DisplayName("Order 생성 검증 시")
+    class ValidateCreateOrder {
 
         @Test
         @DisplayName("OrderTableId에 해당하는 OrderTable이 존재하지 않아서 예외가 발생한다.")
         void throws_notExistOrderTable() {
             // given
             final Long notExistOrderTableId = -1L;
-            final OrderCreateValidationEvent orderCreateValidationEvent = new OrderCreateValidationEvent(notExistOrderTableId);
 
             // when & then
-            assertThatThrownBy(() -> orderTableEventListener.validateOrderTable(orderCreateValidationEvent))
+            assertThatThrownBy(() -> orderValidator.validateCreate(notExistOrderTableId))
                     .isInstanceOf(OrderTableException.NotFoundOrderTableException.class)
                     .hasMessage("[ERROR] 해당하는 OrderTable이 존재하지 않습니다.");
         }
@@ -45,14 +44,12 @@ class OrderTableEventListenerTest extends ServiceTest {
             // given
             final OrderTable orderTable = new OrderTable(ORDER_TABLE1_NUMBER_OF_GUESTS, true);
             final OrderTable savedOrderTable = orderTableRepository.save(orderTable);
-            final OrderCreateValidationEvent orderCreateValidationEvent = new OrderCreateValidationEvent(savedOrderTable.getId());
 
             // when & then
-            assertThatThrownBy(() -> orderTableEventListener.validateOrderTable(orderCreateValidationEvent))
+            assertThatThrownBy(() -> orderValidator.validateCreate(savedOrderTable.getId()))
                     .isInstanceOf(OrderException.CannotOrderStateByOrderTableEmptyException.class)
                     .hasMessage("[ERROR] 주문 테이블이 비어있는 상태일 때 주문할 수 없습니다.");
         }
     }
-
 
 }
